\name{batchColon}
\alias{batchColon}
\docType{data}
\title{
A vector identifying batches in the colon dataset.
}
\description{
A vector identifying batches in the colon dataset.
}
\usage{data("batchColon")}
\format{
  The format is:
 num [1:112] 1 1 1 1 1 1 1 1 1 1 ...
}
\details{
A vector identifying batches in the colon dataset.
}
\source{
Nyamundanda, G., Poudel, P., Patil, Y. and Sadanandam, A., 2017. A novel and robust statistical method to diagnose and correct batch effects in genomic data.
}
\references{
Nyamundanda, G., Poudel, P., Patil, Y. and Sadanandam, A., 2017. A novel and robust statistical method to diagnose and correct batch effects in genomic data.
}
\examples{
data(batchColon)
}
\keyword{datasets}
