\name{bioCL}
\alias{bioCL}
\docType{data}
\title{
A vector identifying tumor and normal samples in the colon dataset.
}
\description{
A vector identifying tumor and normal samples in the colon dataset.
}
\usage{data("bioCL")}
\format{
  The format is:
 num [1:112] 2 2 2 2 2 2 2 2 2 2 ...
}
\details{
A vector identifying tumor and normal samples in the colon dataset.
}
\source{
Nyamundanda, G., Poudel, P., Patil, Y. and Sadanandam, A., 2017. A novel and robust statistical method to diagnose and correct batch effects in genomic data.
}
\references{
Nyamundanda, G., Poudel, P., Patil, Y. and Sadanandam, A., 2017. A novel and robust statistical method to diagnose and correct batch effects in genomic data.
}
\examples{
data(bioCL)

}
\keyword{datasets}
