\name{chol.inv}
\alias{chol.inv}
\title{
Matrix inversion
}
\description{
Matrix inversion 
}
\usage{
chol.inv(x)
}
\arguments{
  \item{x}{
p by p matrix for inverting
}
}
\details{
Matrix inversion
}
\value{
Matrix inversion
}
\references{
Nyamundanda, G., Poudel, P., Patil, Y. and Sadanandam, A., 2017. A novel and robust statistical method to diagnose and correct batch effects in genomic data.
}
\author{
Nyamundanda, G., Poudel, P., Patil, Y. and Sadanandam, A.
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line

