\name{mdiag}
\alias{mdiag}
\title{
Sum diagonal elements of a matrix
}
\description{
Sum diagonal elements of a matrix
}
\usage{
mdiag(x, p)
}
\arguments{
  \item{x}{
P by p matrix
}
  \item{p}{
Number of features
}
}
\details{
Sum diagonal elements of a matrix
}
\value{
Sum diagonal elements of a matrix
}
\references{
Nyamundanda, G., Poudel, P., Patil, Y. and Sadanandam, A., 2017. A novel and robust statistical method to diagnose and correct batch effects in genomic data.
}
\author{
Nyamundanda, G., Poudel, P., Patil, Y. and Sadanandam, A.
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line