\name{pen}
\alias{pen}
\title{
Penalty term for BIC
}
\description{
Penalty term for BIC
}
\usage{
pen(q, p, N)
}
\arguments{
  \item{q}{
Number of PCs
}
  \item{p}{
Number of features
}
  \item{N}{
Number of samples
}
}
\details{
Penalty term for BIC
}
\value{
Penalty term for BIC
}
\references{
Nyamundanda, G., Poudel, P., Patil, Y. and Sadanandam, A., 2017. A novel and robust statistical method to diagnose and correct batch effects in genomic data.
}
\author{
Nyamundanda, G., Poudel, P., Patil, Y. and Sadanandam, A.
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line