\name{ppccajack}
\alias{ppccajack}
\title{
Fitting PPCCA model with jackknifing
}
\description{
Fitting PPCCA model with jackknifing
}
\usage{
ppccajack(Y, res2, n.cores)
}
\arguments{
  \item{Y}{
Expresion data}
  \item{res2}{
Results of PPCCA}
  \item{n.cores}{
Number of corres}
}
\details{
Fitting PPCCA model with jackknifing
}
\references{
Nyamundanda, G., Poudel, P., Patil, Y. and Sadanandam, A., 2017. A novel and robust statistical method to diagnose and correct batch effects in genomic data.
}
\author{
Nyamundanda, G., Poudel, P., Patil, Y. and Sadanandam, A.
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line