\name{scalingfn}
\alias{scalingfn}
\title{
Scaling of data}
\description{
Scaling of data}
\usage{
scalingfn(Y, type)
}
\arguments{
  \item{Y}{
Expression data}
  \item{type}{
Type of scaling}
}
\details{
Scaling of data}
\value{
Scaling of data
}
\references{
Nyamundanda, G., Poudel, P., Patil, Y. and Sadanandam, A., 2017. A novel and robust statistical method to diagnose and correct batch effects in genomic data.
}
\author{
Nyamundanda, G., Poudel, P., Patil, Y. and Sadanandam, A.
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line