\name{unlistfn}
\alias{unlistfn}
\title{
Unlist an lapply function}
\description{
Unlist an lapply function}
\usage{
unlistfn(x, g, byrow)
}
\arguments{
  \item{x}{
object form lapply}
  \item{g}{
number of colums
}
  \item{byrow}{
Fill by row}
}
\details{
Unlist an lapply function}
\value{
Unlist an lapply function
}
\references{
Nyamundanda, G., Poudel, P., Patil, Y. and Sadanandam, A., 2017. A novel and robust statistical method to diagnose and correct batch effects in genomic data.
}
\author{
Nyamundanda, G., Poudel, P., Patil, Y. and Sadanandam, A.
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
